nfRendererOGL4: Implement VBO and IBO
Verification:
    Build, run nfRendererDemo. Scenes 0:1 and 0:2 should now be
    available.

+++ nfEngine/nfCommon/Linux/Window.cpp
@@ -267,7 +267,7 @@ void Window::ProcessMessages()
         case KeyPress:
         {
             mKeys[event.xkey.keycode] = true;
-            OnKeyPress(event.xkey.keycode);
+            OnKeyPress(static_cast<int>(event.xkey.keycode));
             break;
         }
         case KeyRelease:
